sdk/StackReference: Add GetOutputDetailsAsync
This is the C# equivalent of StackReference.GetOutputDetails
that was added to:

- Go SDK in pulumi/pulumi#12034
- Python SDK in pulumi/pulumi#12071
- Node SDK in pulumi/pulumi#12072

It introduces a GetOutputDetailsAsync method
that returns a StackReferenceOutputDetails promise,
allowing looking at the results without going through an Output type.

**Testing**: As with Node (pulumi/pulumi#12072), C# appears to promote
the entire dictionary to secret if one of the values is secret.
Therefore, this needed separate tests for each case.

Refs pulumi/pulumi#10839, pulumi/pulumi#5035