[dotnet/sdk-gen] Fixes potential conflicts when generating resources called System (#14011)
# Description

When generating a C# package that has a resource called `System`, it
fails to compile when we fully qualify
`System.Diagnostics.CodeAnalysis.SuppressMessage` in the generated
`Config.cs` file since it thinks that `System.Diagnostics` is referring
to a static class on the `System` resource definition.

The fix is to prefix the attribute with `global::` so that type
resolution works as expected.

## Checklist

- [ ] I have run `make tidy` to update any new dependencies
- [ ] I have run `make lint` to verify my code passes the lint check
  - [ ] I have formatted my code using `gofumpt`

<!--- Please provide details if the checkbox below is to be left
unchecked. -->
- [ ] I have added tests that prove my fix is effective or that my
feature works
<!--- 
User-facing changes require a CHANGELOG entry.
-->
- [x] I have run `make changelog` and committed the
`changelog/pending/<file>` documenting my change
<!--
If the change(s) in this PR is a modification of an existing call to the
Pulumi Cloud,
then the service should honor older versions of the CLI where this
change would not exist.
You must then bump the API version in
/pkg/backend/httpstate/client/api.go, as well as add
it to the service.
-->
- [ ] Yes, there are changes in this PR that warrants bumping the Pulumi
Cloud API version
<!-- @Pulumi employees: If yes, you must submit corresponding changes in
the service repo. -->