implement debugging support for dotnet
For dotnet to run with a debugger, we don't need to run the binary
under a debugger, but rather the IDEs can just attach to the program.
This needs some support from the SDK, so we wait until the debugger is
actually attached.

Note that this does not seem to work when just running the program
using `dotnet run .`, but rather we need to build and publish the
program first, and only then we can attach the debugger to it.  (It
might just be that running the program under the debugger is horribly
slow, and thus using `dotnet run` takes more time than I had patience
for, but that wouldn't be a good user experience either way).