add Makefile to build the language runtime
Add a simple makefile to build and install the language runtime for
dotnet.  We have a homegrown build system in this repo, but that
doesn't even return the errors for failed Go builds.  Rather than
trying to fix that, it's better to try and converge the build systems
for all our projects, as all others are using Make.