Fix output serialisation in the provider library (#337)
Prompted by reviewing https://github.com/pulumi/pulumi-dotnet/pull/335,
found a number of issues in the serialisation of output values.

This also fixes one of the issues fixed in
https://github.com/pulumi/pulumi-dotnet/pull/335 where `IInput` values
weren't serialised correctly.

To help support testing this I've added a new helper to
`OutputUtilities` to add a dependency to an output. Similar to
`OutputUtilities.CreateUnknown` this isn't expected to be used in normal
programs, but its useful for tests.