implement debugging support for dotnet (#332)
For dotnet to run with a debugger, we don't need to run the binary under
a debugger, but rather the IDEs can just attach to the program. This
needs some support from the SDK, so we wait until the debugger is
actually attached.

Note that this does not seem to work when just running the program using
`dotnet run .`, but rather we need to build and publish the program
first, and only then we can attach the debugger to it. (It might just be
that running the program under the debugger is horribly slow, and thus
using `dotnet run` takes more time than I had patience for, but that
wouldn't be a good user experience either way).

I'm not sure how to write an integration test for this. For Go the
debugger starts a tcp server we can communicate with, but not sure what
attaching to a PID means in that context/how exactly to do that. I did
try it out manually (which was also difficult, since including the local
SDK doesn't seem to work with `dotnet build; dotnet <path-to.dll>`, so I
ended up just copying the SDK changes to a program :grimacing:)

Fixes #333