Allow unknown operations in OperationTypeConverter (#350)
Lessen the strictness of `OperationTypeConverter` to allow unknown
operations. This is a postmortem follow-up after we added a new
operation (`"output-changes"`) in the engine that the .NET Automation
API couldn't handle, resulting in an `InvalidOperationException` being
thrown. Now, when an unknown operation is encountered,
`OperationType.Unknown` will be returned rather than throwing.

Fixes #292