always run all integration tests (#360)
Currently the CI workflow in this repo runs each integration test
individually. This makes it very easy to forget to add a test to the
list, that's then never run in CI, even though it should be.

Fix this by introducing a new Makefile target, to run all integration
tests, and running that in CI instead of spelling each test out
individually, which is very brittle.