[PCL] Support generating invokes options from PCL for invokes (#17696)
### Description

This PR adds support for specifying invokes options in PCL and extends
program-gen in C#, TypeScript, Python and Go to support generating the
options correctly from PCL.

Previously the options parameter from the invoke was expected to be a
string which is incorrect, it must be an object with fields `provider`,
`parent`, `version` and `pluginDownloadUrl`.

Soon we will add `dependsOn` to these options so it will be nice to have
the foundation for it.

For Go, TypeScript and Python added a conformance test
`l2-invoke-options` which exercise the usage of these options although
runtime-wise the results of the invoke are the same.

Since we don't have conformance tests for .NET yet, I've added a simple
program-gen test for C# that has invoke options to make sure the
generated program compiles.