Fix a bug with unknown secret deseralisation (#378)
When deserializing a Secret property value if its inner value was
`Unknown` we would return an input value without `isSecret` set because
we reused a `unknownOutput` variable from higher up.

This removes that variable and replaces the two use sites with calls to
create the output value directly, one without `isSecret` set and one
with it set.