Normalize Computed in OutputReferences to null (#381)
This changes the `PropertyValue` field (`Value`) in `OutputReferences`
to normalize `Computed` values to `null` as we use that to represent
_unknowns_.

As part of this I overhauled the serialisation logic for
`Output<T>`/`Input<T>` to better handle nested property values (e.g. a
secret in an output, or an output in a secret).