Fix warning for Automation API inline programs
When using Automation API inline programs with Pulumi CLI v3.132.0 and
later, a warning is emitted because the language host for .NET inline
programs does not implement the `GetPluginInfo` RPC. This change adds an
implementation for `GetPluginInfo`, which avoids the warning. The
implementation returns "1.0.0" for the version, which matches Node.js
and Go.

Separately, we should consider updating the CLI to not emit the warning
to users when the language host does not implement `GetPluginInfo`.