Fix publishing SDKs to include required attribute for Pulumi.Automation (#393)
Currently when publishing the Pulumi.Automation SDK there is an expected
property PulumiSdkVersion which should be set when building the project

This property is set during buildSdk and testPulumiAutomationSdk but not
the publishing targets which provide the resulting nuget package.

This means when a consumer references the Automation Api nuget package
they have a version that does not have the required property set causing
`LocalPulumiCommand.Install` to fail.

Refactored the F# code to ensure the version is set during publish as
well

Fixes #392