Fix stdout tests
Workaround for https://github.com/dotnet/sdk/issues/44610 In .NET 9.0 `dotnet
run` will print progress indicators to the terminal, even though it should not
do this when the output is redirected.

To allow the tests to continue passing we check for the expected substrings
anywhere in the stdout messages, not just at the start.