Add 'parameterization' to pulumi-plugin.json (#17867)
Part of https://github.com/pulumi/pulumi/issues/17507

For importing resources of parameterized resources we need to be able to
find the parameterized package descriptions without running the users
program. This is the first part of fixing that. We add the parameterized
information about the package into it's pulumi-plugin.json file, so this
file now describes the overall package not just the base provider
plugin.

Once this is merged there are two follow ups;
1) See about fixing sdk-gen to not emit the large byte value twice (in
the source code where we currently have it, and now in the
pulumi-plugin.json). We can probably fix all the sdks to just read from
the embedded plugin file.
2) Update the language hosts to support a "GetAllPackages" instead of
"GetAllPlugins" to read out this information from the pulumi-plugin.json
files.