Fix warning for Automation API inline programs (#388)
When using Automation API inline programs with Pulumi CLI v3.132.0 and
later, a warning is emitted because the language host for .NET inline
programs does not implement the `GetPluginInfo` RPC (from
https://github.com/pulumi/pulumi/pull/17216).

This change adds an implementation for `GetPluginInfo`, which avoids the
warning. The implementation returns `"1.0.0"` for the version, which
matches
[Node.js](https://github.com/pulumi/pulumi/blob/7461bddd65033ba6972e8404b711fd02c1bd5c20/sdk/nodejs/automation/server.ts#L140)
and
[Go](https://github.com/pulumi/pulumi/blob/7461bddd65033ba6972e8404b711fd02c1bd5c20/sdk/go/auto/stack.go#L1596)
(though,
[Python](https://github.com/pulumi/pulumi/blob/7461bddd65033ba6972e8404b711fd02c1bd5c20/sdk/python/lib/pulumi/automation/_server.py#L127-L128)
returns an empty string).

Separately, we should consider updating the CLI to not emit the warning
to users when the language host does not implement `GetPluginInfo`.

Fixes #369