Add Parameterize to the provider interface (#404)
This adds the new Parameterize method to the dotnet provider SDK.

This is tested via a little integration test. I've had to update the C#
testprovider to handle being parameterized, and have written a small
(incomplete) schema by hand to test gen-sdk works with this. This
required renaming the echo resources property because C# can't handle
the resource and property both being called echo. That required updating
some other tests slightly to accommodate the rename.

---------

Co-authored-by: Justin Van Patten <jvp@justinvp.com>