Enable support pack for all parameterized packages (#17923)
SupportPack was added to write out sdks in a new style appropriate for
use in conformance tests and for local sdks. When adding parameterized
we tried to treat them the same, always writing out in the new style.
But we we're inconsistent and didn't always check for "parameterization
!= nil" at the same time as "supportPack".

This changes the schema loader to just do that as part of
marshal/unmarshal schema.

The only thing this actually affects is Go sdkgen in that parameterized
providers now write out their top level go.mod as expected.